Isolate activity grouping from render helpers
